backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
python-socketio>=5.11.0
python-dotenv>=1.0.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-google-genai>=0.0.5
pydantic>=2.5.0
aiosqlite>=0.19.0
websockets>=12.0
requests>=2.31.0

google-genai>=0.3.0
firebase-admin>=6.2.0


# Database Migration
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
greenlet>=3.0.0
